pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "borealis-mcp"
version = "0.1.0"
description = "MCP server for AI agent interaction with ALCF supercomputers"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "ALCF", email = "support@alcf.anl.gov"}
]
keywords = ["mcp", "hpc", "pbs", "alcf", "aurora", "polaris"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "fastmcp>=0.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "mypy>=1.0",
    "ruff>=0.1",
]
http = [
    "requests>=2.28",
]

[project.scripts]
borealis-mcp = "borealis_mcp.server:main"

[project.urls]
Homepage = "https://github.com/argonne-lcf/borealis-mcp"
Documentation = "https://github.com/argonne-lcf/borealis-mcp#readme"
Repository = "https://github.com/argonne-lcf/borealis-mcp"

[tool.hatch.build.targets.wheel]
packages = ["src/borealis_mcp"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
